refactor(files): rewrite to reduce fn parameters
