+++ plugin/src/lib/types.ts
@@ -1,5 +1,6 @@
 import { CollectionConfig, Field, RichTextField } from "payload/types";
 import { type SlateToHtmlConfig, type HtmlToSlateConfig } from '@slate-serializers/html'
+import { CrowdinCollectionDirectory } from "./payload-types";
 
 type CollectionOrGlobalConfigSlug = string
 type CollectionOrGlobalConfigObject = {
@@ -45,3 +46,8 @@ export interface PluginOptions {
 }
 
 export type FieldWithName = Field & { name: string };
+
+// Type checkers
+export const isCrowdinCollectionDirectory = (val: CrowdinCollectionDirectory | string | undefined | null): val is CrowdinCollectionDirectory => {
+  return val !== undefined && val !== null && typeof val !== 'string';
+}
